+++ src/main/php/lang/ast/emit/Type.class.php
@@ -12,6 +12,34 @@ abstract class Type {
   public abstract function name();
 
   /**
+   * Checks whether a given method exists
+   *
+   * @param  string $named
+   * @param  ?int $select
+   * @return bool
+   */
+  public abstract function providesMethod($named, $select= null);
+
+  /**
+   * Checks `#[Override]`
+   *
+   * @param  self $type
+   * @return void
+   * @throws lang.ast.Error
+   */
+  public abstract function checkOverrides($type);
+
+  /**
+   * Checks `#[Override]` for a given method
+   *
+   * @param  string $method
+   * @param  int $line
+   * @return void
+   * @throws lang.ast.Error
+   */
+  public abstract function checkOverride($method, $line);
+
+  /**
    * Returns whether a given member is an enum case
    *
    * @param  string $member