+++ src/test/php/lang/ast/unittest/emit/LambdasTest.class.php
@@ -225,4 +225,36 @@ class LambdasTest extends EmittingTest {
       }
     }');
   }
+
+  #[Test]
+  public function issue_176() {
+    $r= $this->run('class %T {
+      public function run(iterable $records) {
+        $nonNull= fn($record) => null !== $record;
+        $process= fn($records, $filter) => {
+          foreach ($records as $record) {
+            if ($filter($record)) yield $record;
+          }
+        };
+
+        return $process($records, $nonNull);
+      }
+    }', [1, null, 2]);
+
+    Assert::equals([1, 2], iterator_to_array($r));
+  }
+
+  #[Test]
+  public function use_with_return_type() {
+    $r= $this->run('class %T {
+      public function run() {
+        $local= 1;
+        return function() use($local): int {
+          return $local;
+        };
+      }
+    }');
+
+    Assert::equals(1, $r());
+  }
 }